feat(index.html): add initial HTML structure and form for adding todos
This commit adds the initial HTML structure for the todo application. It includes a header with the title "Todo List" and a form for adding new todos. The form includes input fields for the todo title and the due date. There is also a submit button to add the todo. Additionally, a container is added to display the list of todos.