feat(assets): add check-outline.svg, trash-fill.svg, and trash-outline.svg icons
Add three new SVG icons to the assets folder: check-outline.svg, trash-fill.svg, and trash-outline.svg. These icons are used for various UI elements in the application. The icons are in the color #5f30e2 and have a stroke width of 1. They are added in compliance with the design requirements for the application.