feat(script.js): add undoTaskFromCompleted function to allow undoing a completed task
The undoTaskFromCompleted function is added to the script.js file. This function takes a todoId as a parameter and finds the corresponding todo object. If the todo object is found, the isCompleted property is set to false, indicating that the task is no longer completed. Finally, a RENDER_EVENT is dispatched to trigger a re-render of the document. This function allows users to undo a completed task.