feat(script.js): add toast notification to show when data is saved and updated
The toast notification is added to provide a visual feedback to the user when the data is saved and updated. It displays a message indicating that the data has been updated and will automatically disappear after 3 seconds. This improves the user experience by providing real-time feedback on the status of the data.